pyproject.toml
[tool.poetry]
name = "youtube2podbean"
version = "0.1.0"
description = ""
authors = ["Nima Shoghi <nimashoghi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests-oauthlib = "^1.2"
flask = "^1.0"
pafy = "^0.5.4"
youtube-dl = "^2019.5"
discord_webhook = "^0.4.1"
python-dateutil = "^2.8"
colorthief = "^0.2.1"
python-wordpress-xmlrpc = "^2.3"
xmltodict = "^0.12.0"
aiohttp = "^3.5"

[tool.poetry.dev-dependencies]
pylint = {version = "^2.3",allows-prereleases = true}
black = {version = "^18.3-alpha.0",allows-prereleases = true}
bandit = "^1.6"
flake8 = "^3.7"
mypy = "^0.701.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
